# excel-api-node/resources/openapi.yaml
openapi: "3.1.0"
info:
  title: "Excel File API"
  description: "HTTP service exposing Excel spreadsheet data as a JSON API"
  version: "0.0.3"
  contact:
    name: "Filip Golewski"
  license:
    name: "MIT"
servers:
  - url: "http://localhost:8443/api/v1"
    description: "Local development"
security:
  - bearerJwt: []
  - staticToken: []
paths:
...
  /workbooks:
    get:
      summary: "List registered workbooks"
      operationId: "listWorkbooks"
      tags: ["Workbooks"]
      responses:
        "200":
          description: "Workbook list"
          content:
            application/json:
              schema:
                type: object
                required: ["items"]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/WorkbookInfo"
        "429":
          description: "Rate limit exceeded (error code RATE_LIMITED)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /workbooks/{fileId}:
    get:
      summary: "Workbook details"
      operationId: "getWorkbook"
      tags: ["Workbooks"]
      parameters:
        - $ref: "#/components/parameters/fileId"
      responses:
        "200":
          description: "Workbook details"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkbookInfo"
        "404":
          description: "Workbook not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /workbooks/{fileId}/sheets/{sheetName}:
    get:
      summary: "Sheet metadata"
      operationId: "getSheet"
      tags: ["Sheets"]
      parameters:
        - $ref: "#/components/parameters/fileId"
        - $ref: "#/components/parameters/sheetName"
      responses:
        "200":
          description: "Sheet metadata"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SheetInfo"
        "404":
          description: "Workbook or sheet not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /workbooks/{fileId}/sheets/{sheetName}/columns:
    get:
      summary: "Column definitions"
      operationId: "getColumns"
      tags: ["Sheets"]
      parameters:
        - $ref: "#/components/parameters/fileId"
        - $ref: "#/components/parameters/sheetName"
      responses:
        "200":
          description: "Column definitions"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ColumnList"
        "404":
          description: "Workbook or sheet not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /workbooks/{fileId}/sheets/{sheetName}/records:
    get:
      summary: "List records"
      operationId: "listRecords"
      tags: ["Records"]
      parameters:
        - $ref: "#/components/parameters/fileId"
        - $ref: "#/components/parameters/sheetName"
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 1000
        - name: format
          in: query
          schema:
            type: string
            enum: ["native", "display", "string"]
            default: "native"
      responses:
        "200":
          description: "Paginated record list"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordList"
        "400":
          description: "Sheet not configured for tabular access"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: "Add record"
      operationId: "addRecord"
      tags: ["Records"]
      parameters:
        - $ref: "#/components/parameters/fileId"
        - $ref: "#/components/parameters/sheetName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddRecordRequest"
      responses:
        "201":
          description: "Record added"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordItem"
        "409":
          description: "File locked"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: "Readonly workbook"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /workbooks/{fileId}/sheets/{sheetName}/records/{recordIndex}:
    get:
      summary: "Get record"
      operationId: "getRecord"
      tags: ["Records"]
      parameters:
        - $ref: "#/components/parameters/fileId"
        - $ref: "#/components/parameters/sheetName"
        - $ref: "#/components/parameters/recordIndex"
        - name: format
          in: query
          schema:
            type: string
            enum: ["native", "display", "string"]
            default: "native"
      responses:
        "200":
          description: "Record data"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordItem"
        "404":
          description: "Record not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      summary: "Update record"
      operationId: "updateRecord"
      tags: ["Records"]
      parameters:
        - $ref: "#/components/parameters/fileId"
        - $ref: "#/components/parameters/sheetName"
        - $ref: "#/components/parameters/recordIndex"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRecordRequest"
      responses:
        "200":
          description: "Record updated"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordItem"
        "404":
          description: "Record not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: "Delete record"
      operationId: "deleteRecord"
      tags: ["Records"]
      parameters:
        - $ref: "#/components/parameters/fileId"
        - $ref: "#/components/parameters/sheetName"
        - $ref: "#/components/parameters/recordIndex"
      responses:
        "204":
          description: "Record deleted"
        "404":
          description: "Record not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /workbooks/{fileId}/sheets/{sheetName}/operations:
    post:
      summary: "Batch record operations"
      operationId: "batchRecordOperations"
      tags: ["Operations"]
      parameters:
        - $ref: "#/components/parameters/fileId"
        - $ref: "#/components/parameters/sheetName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchRecordRequest"
      responses:
        "200":
          description: "Batch completed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchResult"
        "409":
          description: "File locked"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /workbooks/{fileId}/lock-status:
    get:
      summary: "Lock diagnostic info"
      operationId: "getLockStatus"
      tags: ["Diagnostics"]
      parameters:
        - $ref: "#/components/parameters/fileId"
      responses:
        "200":
          description: "Lock status"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LockStatus"
components:
  securitySchemes:
    bearerJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
    staticToken:
      type: apiKey
      in: header
      name: Authorization
      description: "Format: Token {value}"
    oauth2:
      type: oauth2
      flows:
        password:
          tokenUrl: /auth/token
          scopes:
            read: "Read workbook data"
            write: "Modify workbook data"
            admin: "Administrative operations"
        clientCredentials:
          tokenUrl: /auth/token
          scopes:
            read: "Read workbook data"
            write: "Modify workbook data"
            admin: "Administrative operations"
  parameters:
    fileId:
      name: fileId
      in: path
      required: true
      schema:
        type: string
      description: "Workbook identifier from registry"
    sheetName:
      name: sheetName
      in: path
      required: true
      schema:
        type: string
      description: "Worksheet name (URL-encoded)"
    recordIndex:
      name: recordIndex
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      description: "1-based record index among data rows"
...
  schemas:
...
    WorkbookInfo:
      type: object
      required: ["id", "filename", "readonly", "sheets"]
      properties:
        id:
          type: string
        filename:
          type: string
        readonly:
          type: boolean
        sheets:
          type: array
          items:
            type: string
        modified_at:
          type: string
          format: date-time
        size_bytes:
          type: integer
    SheetInfo:
      type: object
      required: ["name", "row_count", "column_count"]
      properties:
        name:
          type: string
        row_count:
          type: integer
        column_count:
          type: integer
        mode:
          type: string
          enum: ["table", "raw"]
        header_row:
          type: integer
        first_data_row:
          type: integer
    ColumnList:
      type: object
      required: ["source", "columns"]
      properties:
        source:
          type: string
          enum: ["header_row", "multi_row", "legend_sheet"]
        columns:
          type: array
          items:
            $ref: "#/components/schemas/ColumnInfo"
    ColumnInfo:
      type: object
      required: ["index", "letter", "id"]
      properties:
        index:
          type: integer
        letter:
          type: string
        id:
          type: string
        type:
          type: string
          enum: ["string", "number", "date", "boolean", "formula"]
        number_format:
          type: string
          nullable: true
        descriptions:
          type: object
          additionalProperties:
            type: string
    RecordList:
      type: object
      required: ["items", "total", "offset", "limit", "format"]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/RecordItem"
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
        format:
          type: string
          enum: ["native", "display", "string"]
    RecordItem:
      type: object
      required: ["index", "data"]
      properties:
        index:
          type: integer
        data:
          type: object
          additionalProperties: true
        warnings:
          type: array
          items:
            $ref: "#/components/schemas/CellWarning"
    CellWarning:
      type: object
      required: ["column", "code", "message"]
      properties:
        column:
          type: string
        code:
          type: string
        message:
          type: string
    AddRecordRequest:
      type: object
      required: ["data"]
      properties:
        data:
          type: object
          additionalProperties: true
        after_row:
          type: integer
        copy_style_from:
          type: integer
    UpdateRecordRequest:
      type: object
      required: ["data"]
      properties:
        data:
          type: object
          additionalProperties: true
...
    BatchRecordRequest:
      type: object
      required: ["operations"]
      properties:
        operations:
          type: array
          items:
            type: object
            required: ["op"]
            properties:
              op:
                type: string
                enum: ["add", "update", "delete"]
              row_index:
                type: integer
              data:
                type: object
                additionalProperties: true
              copy_style_from:
                type: integer
...
    BatchResult:
      type: object
      required: ["results", "applied_at"]
      properties:
        results:
          type: array
          items:
            type: object
            required: ["op", "status"]
            properties:
              op:
                type: string
              status:
                type: string
              index:
                type: integer
        applied_at:
          type: string
          format: date-time
    LockStatus:
      type: object
      required: ["locked"]
      properties:
        locked:
          type: boolean
        locked_by:
          type: string
        locked_since:
          type: string
          format: date-time
        queue_depth:
          type: integer
    ErrorResponse:
      type: object
      required: ["error", "message"]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true